Updates based on latest C++ Toolbox updates. Minor updates to comments and style. Removed warnings and errors on Windows. Changed CAudioSession::shared to CAudioSession::mShared. Changed CThread proc to take const CThread. Changed TBuffer to return reference instead of value. Added UniversalTime to CPreferences on Windows.

+++ Source/Sources/CTextParceller.cpp
@@ -71,7 +71,7 @@ CString CTextParceller::readStringToEOL(UError& outError)
 			return outString;
 		}
 
-		TBuffer<char>	buffer(bytesRead + 1);
+		TBuffer<char>	buffer((UInt32) bytesRead + 1);
 		outError = mInternals->mDataSource->readData(*buffer, bytesRead);
 		if (outError != kNoError)
 			// Error
